refactor: move monitor swap out method
