+++ daemon/daemon/daemon.go
@@ -178,63 +178,3 @@ func (m *SwapMonitor) MonitorSwapIn(ctx context.Context, currentSwap models.Swap
 
 	return nil
 }
-
-func (m *SwapMonitor) MonitorSwapOut(ctx context.Context, currentSwap models.SwapOut) error {
-	logger := log.WithField("id", currentSwap.SwapID)
-	logger.Info("processing swap out")
-
-	newSwap, err := m.swapClient.GetSwapOut(ctx, currentSwap.SwapID)
-	switch {
-	case errors.Is(err, swaps.ErrSwapNotFound):
-		logger.Warn("swap not found")
-
-		outcome := models.OutcomeFailed
-		currentSwap.Outcome = &outcome
-		currentSwap.Status = models.StatusDone
-
-		err := m.repository.SaveSwapOut(&currentSwap)
-		if err != nil {
-			return fmt.Errorf("failed to save swap out: %w", err)
-		}
-	case err != nil:
-		return fmt.Errorf("failed to get swap out: %w", err)
-	}
-
-	newStatus := models.SwapStatus(newSwap.Status)
-	changed := currentSwap.Status != newStatus
-	switch newStatus {
-	case models.StatusCreated:
-		logger.Debug("Waiting for payment")
-	case models.StatusInvoicePaymentIntentReceived:
-		logger.Debug("Off-chain payment detected")
-	case models.StatusContractFundedUnconfirmed:
-		logger.Debug("On-chain payment detected, waiting for confirmation")
-		currentSwap.TimeoutBlockHeight = int64(newSwap.TimeoutBlockHeight)
-	case models.StatusContractFunded:
-		logger.Debug("Contract funded confirmed, claiming on-chain tx")
-		// TODO: claim the on-chain tx
-		err := m.ClaimSwapOut(ctx, &currentSwap)
-		if err != nil {
-			return fmt.Errorf("failed to claim swap out: %w", err)
-		}
-	case models.StatusContractClaimedUnconfirmed:
-		logger.Debug("40swap has published the claim transaction, waiting for confirmation")
-	case models.StatusDone:
-		// Once it gets to DONE, we update the outcome
-		currentSwap.Outcome = &newSwap.Outcome
-	case models.StatusContractExpired:
-	case models.StatusContractRefundedUnconfirmed:
-	}
-
-	if changed {
-		currentSwap.Status = newStatus
-		err := m.repository.SaveSwapOut(&currentSwap)
-		if err != nil {
-			return fmt.Errorf("failed to save swap out: %w", err)
-		}
-	}
-
-	logger.Infof("swap out processed")
-
-	return nil
-}
